clean up providers
