+++ dapp/src/app/(app)/layout.tsx
@@ -1,13 +1,7 @@
 import { Inter } from 'next/font/google'
-import { cookieToInitialState } from 'wagmi'
-import { headers } from 'next/headers'
-import { TRPCReactProvider } from '~/trpc/react'
-import '~/styles/globals.css'
-import { ClientProviders } from '~/app/_components/client-providers'
-import StyledComponentsRegistry from '~/styles/registry'
-import { wagmiConfig } from '~/lib/wagmi-config'
-import { getServerAuthSession } from '~/server/auth'
 import { SidebarLayout } from '~/app/_components/layouts/SidebarLayout'
+import { Providers } from '~/app/_components/providers'
+import '~/styles/globals.css'
 
 const inter = Inter({
 	subsets: ['latin'],
@@ -31,21 +25,12 @@ export const metadata = {
 }
 
 export default async function RootLayout({ children }: { children: React.ReactNode }) {
-	const initialState = cookieToInitialState(wagmiConfig, headers().get('cookie'))
-	const session = await getServerAuthSession()
-
 	return (
 		<html lang='en'>
 			<body className={`font-sans ${inter.variable}`} style={{ margin: 0 }}>
-				<TRPCReactProvider>
-					<StyledComponentsRegistry>
-						<ClientProviders initialState={initialState} session={session}>
-						<SidebarLayout>
-							{children}
-						</SidebarLayout>
-						</ClientProviders>
-					</StyledComponentsRegistry>
-				</TRPCReactProvider>
+				<Providers>
+					<SidebarLayout>{children}</SidebarLayout>
+				</Providers>
 			</body>
 		</html>
 	)
